Move logged entities from Region to EntityData
