+++ src/main/java/com/shanebeestudios/hg/game/GameEntityData.java
@@ -2,8 +2,14 @@ package com.shanebeestudios.hg.game;
 
 import com.shanebeestudios.hg.data.MobData;
 import com.shanebeestudios.hg.data.MobEntry;
+import com.shanebeestudios.hg.plugin.HungerGames;
 import org.bukkit.Location;
 import org.bukkit.entity.Entity;
+import org.bukkit.metadata.FixedMetadataValue;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * Data holder for a {@link Game Game's} entities
@@ -11,7 +17,10 @@ import org.bukkit.entity.Entity;
 @SuppressWarnings("unused")
 public class GameEntityData extends Data {
 
+    private static final FixedMetadataValue SPAWN_KEY = new FixedMetadataValue(HungerGames.getPlugin(), true);
+
     private MobData mobData;
+    private final List<Entity> entities = new ArrayList<>();
 
     GameEntityData(Game game) {
         super(game);
@@ -52,11 +61,69 @@ public class GameEntityData extends Data {
         if (mobEntry != null) {
             Entity spawn = mobEntry.spawn(location);
             if (spawn != null) {
-                this.game.getGameArenaData().getGameRegion().addEntity(spawn);
+                logEntity(spawn);
                 return true;
             }
         }
         return false;
+    }
+
+    /**
+     * Kill/Remove all entities in the arena
+     */
+    public void removeEntities() {
+        List<Entity> entitiesToRemove = new ArrayList<>(this.entities);
+        entitiesToRemove.forEach(Entity::remove);
+        this.entities.clear();
+    }
+
+    /**
+     * Remove a logged entity
+     *
+     * @param entity Entity to remove
+     */
+    public void removeEntityFromLog(Entity entity) {
+        this.entities.remove(entity);
+    }
+
+    /**
+     * Log an entity to be removed later
+     *
+     * @param entity Entity to log
+     */
+    public void logEntity(@NotNull Entity entity) {
+        if (this.entities.contains(entity)) return;
+        entity.setPersistent(false);
+        entity.setMetadata("hunger-games-spawned", SPAWN_KEY);
+        this.entities.add(entity);
+    }
+
+    /**
+     * Check if this entity is already logged
+     *
+     * @param entity Entity to check
+     * @return True if entity is already logged
+     */
+    public boolean hasLoggedEntity(Entity entity) {
+        return this.entities.contains(entity);
+    }
+
+    /**
+     * Get a list of all entities logged
+     *
+     * @return Entities logged
+     */
+    public List<Entity> getLoggedEntities() {
+        return this.entities;
+    }
+
+    /**
+     * Get amount of entities logged
+     *
+     * @return Count of entities logged
+     */
+    public int getLoggedEntityCount() {
+        return this.entities.size();
     }
 
 }
